restructured project directory tree
