+++ back/src/main/java/com/ares_expedition/controller/api/HelloController.java
@@ -1,4 +1,4 @@
-package com.ares_expedition.api.controller;
+package com.ares_expedition.controller.api;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
