+++ back/src/main/java/com/ares_expedition/repository/GameController.java
@@ -1,4 +1,4 @@
-package com.ares_expedition.database;
+package com.ares_expedition.repository;
 
 import java.util.ArrayList;
 import java.util.List;
